[TASK] Move ext_icon.png to Resources/Public/Icons for all sysexts
All extension icons of all system and test extensions shipped with
the core are now located at EXT:.../Resources/Public/Icons/Extension.*
instead of ext_icon.gif.

This is one further step to make it possible to publish only public
assets of an extension to the htdocs/ folder.

Resolves: #78334
Releases: master
Change-Id: Ic65036bb74d4a156c1a0a9fdfa0c8d4b5a28e9ed
Reviewed-on: https://review.typo3.org/50233
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>