[FEATURE] TCA: Allow label in palette array
Allow 'label' within the 'palettes' array in TCA next
to 'showitem'. This defines a default label rendered as
the palette label in FormEngine if not overridden in
a 'types' definition like '--palette--;labelReference;thePalette'.

The patch introduces this feature in PaletteAndSingleContainer
of FormEngine code and adapts core TCA definitions to introduce
the default labels for palettes.

This makes the 'types' definitions shorter and much better readable
and gives the core an opportunity to steer the labels for palettes
which are often re-used by extension authors.

Change-Id: I185340d4e037d713eebbb6cd676287e6fe85705b
Resolves: #82826
Releases: master
Reviewed-on: https://review.typo3.org/54449
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>