[TASK] Merge saltedpasswords felogin hook into felogin
Extension salted passwords hooks into felogin to hash a new
password before it is written to database if the user resets
his password using the 'forgot password' functionality.
Since salted passwords is a mandatory extension, this hook
usage is now removed and the hash creation is put into
felogin directly.
Note the hook method feloginForgotPasswordHook() had a type
hint to FrontendLoginController, is thus tailored for this
extension only and can't be abused by a different one. It is
safe to remove that method entirely.

Change-Id: I85d4ba59cecd3bd43b148008e70fe20f0f5dc0bc
Resolves: #85703
Releases: master
Reviewed-on: https://review.typo3.org/57752
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>