[FEATURE] Extbase-based Frontend Login Form
A new Extbase-based plugin is added to TYPO3's Extension "felogin"
which can be used as a toggle with custom templates based on Fluid,
instead of marker-based templates.

A new feature toggle is added to the Settings module for Site Administrators
to switch to the newly added felogin extbase plugin, or continuing using
the pibase plugin, e.g. for upgrading purposes.

Resolves: #84262
Releases: master
Change-Id: I9d281912373a078e0403f52b27483dd3e04785f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61900
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Alexander Schnitzler <review.typo3.org@alexanderschnitzler.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Alexander Schnitzler <review.typo3.org@alexanderschnitzler.de>
Reviewed-by: Susanne Moog <look@susi.dev>