[TASK] Streamline felogin TypoScript and documentation
Since the removal of the old pi-based plugin some of the
old TypoScript constants in styles.content.loginform
were left unassigned.

With this patch, all TypoScript constants are assigned
to their associated setup variables, so users updating
from the pi-based plugin and mainly using constants for
configuration can easily migrate to the Extbase version
of the plugin.

The following unused TypoScript constants are removed:

* templateFile
* feloginBaseURL

The constants `pid` and `recursive` are mapped to the
TypoScript setup `pages` and `recursive`.

The TypoScript setup variable `replyTo` is removed,
since it is unused in the extension and the settings
has been replaced previously by
`TYPO3_CONF_VARS.MAIL.defaultMailReplyToAddress`

The documentation is adapted to match the current
state of the extension and existing images have been
replaced with current versions.

Resolves: #94328
Releases: master
Change-Id: I509201b1e2ce1194e70af3a618b42ba6c5d80ce3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69471
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>