[BUGFIX] Fix EXT:felogin FlexForm
EXT:felogin uses checkbox items in their FlexForm
configuration. Since the item key does not start
with 0 but with 1, it's no longer possible to save
the items. This previously only worked because
of another bug in the checkbox element, which
however had been fixed in #95371.

This is now fixed by correctly using 0 as the
first item key for the checkbox items.

Background: The item key is used for the
"is checked" calculation wich uses arithmetic
operators and requires the first item key
to be 0.

Resolves: #95510
Related: #95371
Releases: master
Change-Id: Icb71d5b8d8beeba9f6b11bb03f72d45fbbdedb3e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71499
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>