[TASK] Improve sysext metadata
The title and description texts as well as the README.rst for
system extensions are unified and some texts updated to
reflect changes.

Since the description and README.rst are also displayed on
Packagist, the text should be chosen so that it helps
people not familiar with TYPO3 know what the package is
about (e.g. by preceding the description with "TYPO3 CMS").

Descriptive texts for extensions are in these files:

- composer.json
- ext_emconf.php
- README.rst

The information is displayed in the Extension Manager (ext_emconf.php),
on extensions.typo3.org, on Packagist (composer.json and README.rst)
and on GitHub (README.rst).

We reuse the same text blocks in the following manner:

1. Extension title - used in ext_emconf.php:title and composer.json:
   description
2. Extension description (short) - used in ext_emconf.php:description
   and composer.json:description
3. Title and extension description (long) - used in README.rst. In
   most cases, the short description is reused, in some cases it is
   augmented with more information.

In composer.json, the description consists of: '<title> - <description>'

Also,

- In the title all words are capitalized (common name)
- The title always starts with "TYPO3 CMS"
- The description ends with a period.

Resolves: #96059
Releases: master
Change-Id: I57388ff96c5fcf9a8e64cd86ed8427f70d7ef23d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72288
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>