[!!!][BUGFIX] Set password forgot hash based on user uid in ext:felogin
When the forgot password function is used, TYPO3 looks up the affected
frontend user record by either the email address or the username. If
a user record is found, the email address of that record is used to
set the password forgot hash. This may lead to the situation, that
the password forgot hash is set to multiple frontend user records
containing the same email address (e.g. in different storage PIDs).

With this change, the calculated password forgot hash is now saved
based on the frontend users uid.

Additionally, the sendRecoveryEmail function in RecoveryService has
been streamlined and simplified. The superfluous and wrong lookup
of the frontend user by the email address in RecoveryService has
been removed and the function signature of sendRecoveryEmail has
been changed to support an array of given userdata and a
pre-calculated password forgot hash.

Resolves: #95132
Related: #97784
Releases: main
Change-Id: I32c3256ef88649e5bc0b4fa660eb780cc01e6389
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75274
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>