[TASK] Ensure to have a default showitem info when adding plugins as CType
This change adds the basic information for a type
when adding a plugin as CType. When adding a plugin with
"list_type" (default), then the "list" CType is used,
and extension authors are able to add or remove fields
easily with the API. However, when encouraging extension authors
to use CType, this is currently much more boilerplate and not
possible to make this more flexible.

As shown, EXT:felogin and EXT:form previously used this concept
manually, which does not allow to extend fields properly,
but extensions with plugins as CTypes had to define them explicitly.

As TYPO3 Core encourages people to use custom CTypes in favor
of list_type (moving towards a unifieid Content Block concept),
this change will also be available for 11.5.

Resolves: #98960
Releases: main, 11.5
Change-Id: Iee5e1a35f2d99faf03081d7265c2fe6e4845397f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76319
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>