[TASK] Optimize referer evaluation in ext:felogin
The redirect evaluation is currently always processed no matter if
a `redirectMode` based on the referer is configured or not. Also, given
referers are checked for validity using `RedirectUrlValidator`, which
logs a message in the TYPO3 log with the log level `warning` if the given
URL is considered as invalid.

This patch adds a check to the referer evaluation, so it is only processed,
if a redirect mode supporting the referer is active. Additionally, the log
level for invalid redirect URLs in `RedirectUrlValidator` is changed to
`debug`, since a warning is not appropriate for invalid redirect URLs.

Resolves: #100197
Releases: main, 11.5
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I10d1723de131e0387f7dd2103c8a1ca21ed015e5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78169
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>