[BUGFIX] Handle https in url for felogin redirect mode 'refererDomains'
The `redirectMode=refererDomains` configuration allows to process a
redirect after a successful login to an evaluated referer value.

This patch fixes the evaluation of https urls as referer value. Without
this patch only http urls are evaluated correctly.

Also do not match hostnames that contain an invalid underscore in name.

Additionally the regex matching for the domain was not quoting the
configured domain names and thereby matching dots as 'any single
character' instead of a dot.

Resolves: #100215
Releases: main, 12.4, 11.5
Change-Id: I589638d82627283b5acdeec5da0243e082cc5ebf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80132
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>