[TASK] Simplify fe_login template by removing unused onSubmit
Replace the Fluid if condition for the onSubmit value in the login template of EXT:fe_login with the else part.

The onSubmit value in the login template of EXT:fe_login was
ported from the pibase'd version but never set in Extbase'd
version and thus in Fluid.

Resolves: #101178
Releases: main, 12.4
Change-Id: Ic80a65c80841292c4b6ee2251d8e148d51b56e9c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79567
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>