[TASK] Remove superfluous code in ext:felogin
With #98370, all occurrences of `setOriginalRequestMappingResults()`
have been replaced with the reassignment of the global Extbase
request object attached with a modified `Result` object.

This is however not required and has no effect, since in all
cases, an Extbase `ForwardResponse` is returned, which includes
the modified `Result` object. The `Result` object is then used
in Extbase Dispatcher to build the new Extbase request object,
which then is redispatched.

This change therefore removes the superfluous Extbase request
object assignments in `PasswordRecoveryController`.

Additionally, the variable name for the original result has
been streamlined to `$originalResult`.

Resolves: #102112
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I28cdc69d2772dc87543a39d13305be610ee94a54
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81372
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>