[TASK] Cleanup PasswordRecoveryController in ext:felogin
Some method names and descriptions in PasswordRecoveryController are not
fully correct and should be optimized in order to make clear, what those
functions actually do.

Additionally, the only functionality of the `notifyPasswordChange`
function is to dispatch the `PasswordChangeEvent` PSR-14 event.
The user lookup in `notifyPasswordChange` is superfluous, since the
previously called `validateHashAndPasswords` function did already
verify, that a frontend user with the given password recovery hash
exists.

This change renames the function names of `validateIfHashHasExpired` and
`hasValidHash` to a more suitable name and specifies some doc comments
or comments in the controller. The `notifyPasswordChange` function
has been removed and the `PasswordChangeEvent` PSR-14 event is now
directly dispatched in the `changePasswordAction`.

Resolves: #102280
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I45195ce9d0547e46c3a34d2794d7da1a59f069f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81603
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>