[FEATURE] Provide full userdata in password recovery email
Currently, the variables `receiverName`, `url` and `validUntil` are
available in the template of the password recovery email. If more
user related data is required in the recovery email, integrators
have to use the PSR-14 event `SendRecoveryEmailEvent` to add
additional variables to the email object.

To simplify customization of the password recovery email for
frontend users, this patch adds the new variable `userData` to
the FluidEmail object, which contains all userdata of the
affected frontend user.

Resolves: #100268
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I032f775e01003a5a276f8e2d0310925ea1592dc7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81605
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Benni Mack <benni@typo3.org>