[DOCS] Added note to ext:felogin redirect modes
The documentation for redirect mode "Defined by Domain entries"
for ext:felogin does not describe, that configured domains must
be known to TYPO3.

This change updates the documentation, so it is clear, that all
domains for the "Defined by Domain entries" redirect mode must be
configured in TYPO3 site settings.

Resolves: #98372
Releases: main, 12.4
Change-Id: I2a18ce40016d60c3878656cc5194cec6af5ed6a5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82533
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>