[FEATURE] Introduce ModifyRedirectUrlValidationResultEvent PSR-14 event
The frontend login extension currently validates a given redirect URL
by ensuring, that the URL is either relative or the host is known to
TYPO3. In some scenarios, it might however be useful to allow a
redirect URL, which does not match the existing rules.

This change adds the `ModifyRedirectUrlValidationResultEvent` PSR-14
event, allowing developers to implement custom validation logic for
the redirect URL.

Resolves: #103019
Releases: main
Change-Id: I81b387f23e90adbaf4a586c9108766275f5e1273
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82763
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>