[BUGFIX] Evaluate `noredirect` parameter in ext:felogin
The legacy version of the felogin plugin did evaluate the GET/POST
parameter `noredirect` to disable redirect handling for successful
login and logout. With the refactoring to extbase, the parameter
evaluation has been implemented wrong, since it is checked for
the request argument `noredirect`, which however is never set or
used.

This patch fixes the problem by evaluating GET and POST parameter
`noredirect` in the `isRedirectDisabled` function.

Additionally, a new chapter has been added to the documentation,
which describes the parameter and its behavior. A follow-up
patch will document all other GET and POST parameters the
extension uses.

Resolves: #104398
Releases: main, 12.4
Change-Id: I44afe0fe9f925a945da9a450bc9a97e28e336809
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85313
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>