[!!!][TASK] Do not perform redirect in ext:felogin logoutAction
When `settings.showLogoutFormAfterLogin` is configured in the
ext:felogin plugin, no configured redirect must be processed.
This does however not work, because the `logoutAction`
actually processes a possible configured redirect URL,
which is wrong.

Since the main function of the `logoutAction` is to display
the logout form, the action should never process a possible
configured redirect (e.g. via plugin or GET parameter).
Additionally, redirects are already properly handled in
the `loginAction` (with #103910) and the `overviewAction`.

With this change, the wrong redirect handling in the
`logoutAction` has been removed.

Resolves: #103913
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Iac3d14a0cfa8b825787f4d7541c1d47498c72e72
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88423
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>