[TASK] Add rate limit to ext:felogin password recovery
The password reset functionality in ext:felogin allows
unlimited password reset requests for a given username
or email address. This can be used by a remote user
to flood a TYPO3 user's inbox with password recovery
emails. Additionally, this can result in service
degradation or email server overload. It is however
not possible for a remote user to verify, that a user
exists and if an email is actually sent. The TYPO3
Security Team therefore decided to handle the issue
in public.

This change adds a rate limit to the password recovery
function in ext:felogin. The rate limit is configured
to allow a maximum of 5 password recovery attempts for
a frontend user in a timeframe of 15 minutes.
Additionally, a random wait time (200ms - 3 seconds)
after each reset request is added to prevent possible
time-based information disclosure.

Resolves: #105058
Releases: main, 13.4, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: If3e0c763b7586aa53b93199f1a3fa3e0573c1eea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89035
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>